requirements.txt
lxml==5.3.0
psutil==5.9.0
py==1.11.0
pyshark==0.4.5
PyYAML==6.0.2
